Improve accuracy of STL format detection
Improved the check that determines wether a STL is ASCII or binary. It used to check if the file starts with "solid", which causes some binary files to be wrongly interpreted as ASCII files.
Now it compares the Filesize to the Size that would be expected if it was a Binary file.

Correcting for PR comments